revert(app): drop the creation stopwatch — benchmark stays out of the repo
Timing is measured ad-hoc against the gateway instead; the shipped change
is only the instant setup-mission hello.
